Add `iris bloqs goals` CLI command for business goal management
Streamlines heartbeat goal setting from the CLI — view, add, remove, clear
goals and set business context (industry, audience, summary) on any bloq.
Fixes number_format type error in monitor agent command.

Usage:
  iris bloqs goals <id>                                # View goals
  iris bloqs goals <id> --add "Goal text"              # Add goal(s)
  iris bloqs goals <id> --remove 2                     # Remove by number
  iris bloqs goals <id> --industry "AI" --audience "..." # Set context

+++ src/Console/Commands/MonitorCommand.php
@@ -258,7 +258,7 @@ class MonitorCommand extends Command
                 ['Completed', '<fg=green>' . ($stats['completed'] ?? 0) . '</>'],
                 ['Failed', $failCount > 0 ? '<fg=red>' . $failCount . '</>' : '0'],
                 ['Success Rate', $successRate],
-                ['Total Tokens', number_format($stats['total_tokens'] ?? 0)],
+                ['Total Tokens', number_format((int) ($stats['total_tokens'] ?? 0))],
                 ['Active Jobs', $stats['active_jobs'] ?? 0],
             ];
